+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -7,7 +7,9 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.CartRepositoryImpl
 import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.ProductRepositoryImpl
 import woowacourse.shopping.model.Product
 
 class MainViewModel(
@@ -34,8 +36,8 @@ class MainViewModel(
         val Factory: ViewModelProvider.Factory = viewModelFactory {
             initializer {
                 MainViewModel(
-                    ProductRepository(),
-                    CartRepository(),
+                    ProductRepositoryImpl(),
+                    CartRepositoryImpl(),
                 )
             }
         }
